requirements.txt
fastapi
uvicorn
Jinja2
tinydb
python-multipart
python-dateutil
deta[async]==1.1.0a1